Refactor: Move tests to proper file
Leftover from extraction of Action from Item"
